Move param_resolution related code to package directory and rename linear_two_pass function
