+++ src/multiview_stitcher/param_resolution/global_optimization.py
@@ -5,261 +5,13 @@ import networkx as nx
 import numpy as np
 import pandas as pd
 import xarray as xr
-from skimage.transform import (
-    EuclideanTransform,
-    SimilarityTransform,
-)
+from skimage.transform import EuclideanTransform, SimilarityTransform
 
 from multiview_stitcher import mv_graph, param_utils, transformation
-from multiview_stitcher.param_resolution_utils import (
-    compute_edge_residuals,
-    get_beads_graph_from_reg_graph,
-    get_graph_ndim,
-)
-from multiview_stitcher.linear_two_pass import (
-    linear_two_pass_groupwise_resolution,
-)
 from multiview_stitcher.transforms import AffineTransform, TranslationTransform
+from .utils import get_beads_graph_from_reg_graph, get_graph_ndim
 
 logger = logging.getLogger(__name__)
-
-_GROUPWISE_RESOLUTION_METHODS = {}
-
-
-def register_groupwise_resolution_method(name, resolver):
-    """
-    Register a groupwise resolution method.
-
-    The resolver is called once per connected component and must implement:
-    resolver(g_reg_component_tp, **kwargs) -> (params, info_dict).
-    The input graph contains single-timepoint transforms.
-    """
-
-    if not callable(resolver):
-        raise TypeError("Resolver must be callable.")
-    _GROUPWISE_RESOLUTION_METHODS[name] = resolver
-
-
-def _get_groupwise_resolution_method(method):
-    if callable(method):
-        return method
-    if method in _GROUPWISE_RESOLUTION_METHODS:
-        return _GROUPWISE_RESOLUTION_METHODS[method]
-    raise ValueError(f"Unknown groupwise optimization method: {method}")
-
-
-def _get_graph_timepoints(g_reg):
-    t_coords = []
-    for e in g_reg.edges:
-        transform = g_reg.edges[e].get("transform")
-        if isinstance(transform, xr.DataArray) and "t" in transform.coords:
-            t_coords.extend(list(transform.coords["t"].values))
-    return sorted(set(t_coords))
-
-
-def groupwise_resolution(g_reg, method="global_optimization", **kwargs):
-    """
-    Resolve global parameters by running a method per connected component
-    and timepoint.
-
-    Parameters
-    ----------
-    method : str or Callable
-        Name of a registered method, or a callable implementing the
-        component-level, single-timepoint resolver API.
-    """
-
-    resolver = _get_groupwise_resolution_method(method)
-    if not len(g_reg.edges):
-        raise (
-            mv_graph.NotEnoughOverlapError(
-                "Not enough overlap between views\
-        for stitching."
-            )
-        )
-
-    # if only two views, set reference view to the first view
-    # this is compatible with a [fixed, moving] convention
-    if "reference_view" not in kwargs and len(g_reg.nodes) == 2:
-        kwargs["reference_view"] = min(list(g_reg.nodes))
-
-    params = {node: [] for node in g_reg.nodes}
-    info_metrics = []
-    used_edges_by_t = {}
-
-    # Resolve per timepoint and connected component, then stitch results.
-    t_coords = _get_graph_timepoints(g_reg)
-
-    # Normalize to a single-timepoint loop for uniform handling.
-    iter_t_coords = t_coords if t_coords else [None]
-    for it, t in enumerate(iter_t_coords):
-        g_reg_t = (
-            get_reg_graph_with_single_tp_transforms(g_reg, t)
-            if t is not None
-            else g_reg
-        )
-        for icc, cc in enumerate(nx.connected_components(g_reg_t)):
-            g_reg_subgraph = g_reg_t.subgraph(list(cc))
-            if not g_reg_subgraph.number_of_edges():
-                ndim = get_graph_ndim(g_reg_subgraph)
-                cc_params = {
-                    node: param_utils.identity_transform(ndim)
-                    for node in cc
-                }
-                cc_info = None
-            else:
-                cc_params, cc_info = resolver(g_reg_subgraph, **kwargs)
-            for node in cc:
-                params[node].append(cc_params[node])
-
-            if cc_info is not None:
-                # Accumulate metrics and edge usage from the resolver.
-                metrics = cc_info.get("metrics")
-                if metrics is not None:
-                    metrics = metrics.copy()
-                    if t is not None:
-                        metrics["t"] = [t] * len(metrics)
-                    if "icc" not in metrics.columns:
-                        metrics["icc"] = [icc] * len(metrics)
-                    info_metrics.append(metrics)
-                used_edges = cc_info.get("used_edges")
-                if used_edges is not None:
-                    used_edges_by_t.setdefault(it, set()).update(
-                        tuple(sorted(e)) for e in used_edges
-                    )
-
-    # Concatenate per-timepoint parameters.
-    if t_coords:
-        params = {
-            node: xr.concat(params[node], dim="t").assign_coords(
-                {"t": t_coords}
-            )
-            for node in params
-        }
-    else:
-        params = {node: params[node][0] for node in params}
-
-    edge_residuals_by_t = {}
-    for it, t in enumerate(iter_t_coords):
-        params_t = {
-            node: (
-                params[node].sel({"t": t_coords[it]})
-                if isinstance(params[node], xr.DataArray)
-                and "t" in params[node]
-                else params[node]
-            )
-            for node in params
-        }
-        g_reg_t = (
-            get_reg_graph_with_single_tp_transforms(g_reg, t)
-            if t is not None
-            else g_reg
-        )
-        edge_residuals_by_t[it] = compute_edge_residuals(
-            g_reg_t, params_t
-        )
-
-    info_dict = {
-        "metrics": pd.concat(info_metrics) if info_metrics else None,
-        "edge_residuals": edge_residuals_by_t,
-        "used_edges": {k: list(v) for k, v in used_edges_by_t.items()},
-    }
-    return params, info_dict
-
-
-def groupwise_resolution_shortest_paths(g_reg, reference_view=None):
-    """
-    Get final transform parameters by concatenating transforms
-    along paths of pairwise affine transformations.
-
-    Output parameters P for each view map coordinates in the view
-    into the coordinates of a new coordinate system.
-
-    Note: This function operates on a single connected component
-    with single-timepoint transforms.
-    """
-
-    if not g_reg.number_of_edges():
-        ndim = get_graph_ndim(g_reg)
-        params = {
-            node: param_utils.identity_transform(ndim) for node in g_reg.nodes
-        }
-        return params, {"metrics": None, "used_edges": [], "edge_residuals": {}}
-
-    ndim = get_graph_ndim(g_reg)
-
-    # use quality as weight in shortest path (mean over tp currently)
-    # make sure that quality is non-negative (shortest path algo requires this)
-    quality_min = np.min([g_reg.edges[e]["quality"] for e in g_reg.edges])
-    for e in g_reg.edges:
-        g_reg.edges[e]["quality_mean"] = np.mean(g_reg.edges[e]["quality"])
-        g_reg.edges[e]["quality_mean_inv"] = 1 / (
-            (g_reg.edges[e]["quality_mean"] - quality_min) + 0.5
-        )
-
-    # get directed graph and invert transforms along edges
-
-    g_reg_di = g_reg.to_directed()
-    for e in g_reg.edges:
-        sorted_e = tuple(sorted(e))
-        g_reg_di.edges[(sorted_e[1], sorted_e[0])][
-            "transform"
-        ] = param_utils.invert_xparams(g_reg.edges[sorted_e]["transform"])
-
-    node_transforms = {}
-
-    subgraph = g_reg_di
-
-    if reference_view is not None and reference_view in subgraph.nodes:
-        ref_node = reference_view
-    else:
-        ref_node = mv_graph.get_node_with_maximal_edge_weight_sum_from_graph(
-            subgraph, weight_key="quality"
-        )
-
-    # get shortest paths to ref_node
-    paths = {
-        n: nx.shortest_path(
-            subgraph, target=n, source=ref_node, weight="quality_mean_inv"
-        )
-        for n in subgraph.nodes
-    }
-
-    # Track edges that contribute to any shortest path.
-    used_edges = set()
-
-    for n in subgraph.nodes:
-        reg_path = paths[n]
-
-        path_pairs = [
-            [reg_path[i], reg_path[i + 1]]
-            for i in range(len(reg_path) - 1)
-        ]
-        for pair in path_pairs:
-            used_edges.add(tuple(sorted(pair)))
-
-        path_params = param_utils.identity_transform(ndim)
-
-        for pair in path_pairs:
-            path_params = param_utils.rebase_affine(
-                g_reg_di.edges[(pair[0], pair[1])]["transform"],
-                path_params,
-            )
-
-        node_transforms[n] = param_utils.invert_xparams(path_params)
-
-    # homogenize dims and coords in node_transforms
-    # e.g. if some node's params are missing 't' dimension, add it
-    node_transforms = xr.Dataset(data_vars=node_transforms).to_array("node")
-    node_transforms = {
-        node: node_transforms.sel({"node": node}).drop_vars("node")
-        for node in node_transforms.coords["node"].values
-    }
-
-    return node_transforms, {
-        "metrics": None,
-        "used_edges": list(used_edges),
-    }
 
 
 def groupwise_resolution_global_optimization(
@@ -409,26 +161,6 @@ def groupwise_resolution_global_optimization(
     }
 
     return params, info_dict
-
-
-register_groupwise_resolution_method(
-    "global_optimization", groupwise_resolution_global_optimization
-)
-register_groupwise_resolution_method(
-    "shortest_paths", groupwise_resolution_shortest_paths
-)
-register_groupwise_resolution_method(
-    "linear_two_pass", linear_two_pass_groupwise_resolution
-)
-
-
-def get_reg_graph_with_single_tp_transforms(g_reg, t):
-    g_reg_t = g_reg.copy()
-    for e in g_reg_t.edges:
-        for k, v in g_reg_t.edges[e].items():
-            if isinstance(v, xr.DataArray) and "t" in v.coords:
-                g_reg_t.edges[e][k] = g_reg_t.edges[e][k].sel({"t": t})
-    return g_reg_t
 
 
 def optimize_bead_subgraph(
